Add "PSL", mp4 support, and refactoring
The music show "PSL" did not show up in the other listings, added it as separate category.
The show consists of "clips" only, which needs a slightly different logic than the others.

Old "clips" are only available in MP4 format. Added support for this in the video parser.

Refactored the code since there were a lot of duplicated code, and to make it easier to incorporate "PSL".