Add search feature to the plugin
This commit adds the ability to search for an item on SVT Play.
Search results are grouped after type (like on the web page) so that paging can be supported.
This commit also introduces some refactoring for the sake of removing code duplication.
Finally, "PSL" was removed since it can now be found by searching for it.