Fixed listing of episodes.
SVT decided to change their page layout once again. Fetch the JSON for
each episode so that we can get all the needed data. Plot and the URL is
the only thing that needs to be parsed out of the HTML now.

Cleaned up helper functions for the JSON provided data.