The great rewrite
The reasons behind this rewrite is that much of the code is
from when the plugin used to scrape the website.
Since the transition to using APIs the most of the changes has
been more or less hacks.

This mononlith commit contains new code, clean up and removal of
dead code.