Add test for class SvtPlay
This test is supposed to test a real-life scenario for the addon.

In the scenario the user views the A-Z listing and selects one of the items.

It is not determninistic per se as it randomizes what item the user selects.

The purpose of the test is to find irregularities as this scenario
is similar to what how many bugs are reported.